@@ -2,16 +2,46 @@
   "name": "L7ICCSP",
   "version": "1.0.0",
   "description": "Layer 7 Integrity Check for Cloud Service Providers",
-  "main": "index.js",
+  "main": "build/main/main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "electron ."
+    "build-tsc": "tsc",
+    "render": "webpack --config webpack.common.js",
+    "start": "npm run build-tsc && electron .",
+    "dist": "electron-builder"
   },
   "keywords": [],
-  "author": "Siva",
+  "author": "PW21HBP02",
   "license": "ISC",
   "devDependencies": {
-    "electron": "^11.1.1",
-    "typescript": "^4.1.3"
+    "@babel/core": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-react": "^7.12.10",
+    "@types/node": "^14.14.20",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "babel-loader": "^8.2.2",
+    "electron": "^11.2.0",
+    "electron-builder": "^22.9.1",
+    "fork-ts-checker-webpack-plugin": "^6.1.0",
+    "ts-loader": "^8.0.14",
+    "typescript": "^4.1.3",
+    "webpack": "^5.13.0",
+    "webpack-cli": "^4.3.1"
+  },
+  "dependencies": {
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1"
+  },
+  "build": {
+    "appId": "com.app.soft",
+    "files": [
+      "build/**/*",
+      "package.json",
+      "index.html"
+    ],
+    "win": {
+      "target": "nsis"
+    }
   }
 }
